refactor: 세션보드 엔티티 디렉토리 구조 수정
