--- common/src/main/java/com/study/common/entity/SessionSpeaker.java
@@ -1,46 +1,0 @@
-package com.study.common.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import jakarta.persistence.UniqueConstraint;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(
-    name = "session_speaker",
-    uniqueConstraints = @UniqueConstraint(columnNames = {"session_id", "user_id"}))
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SessionSpeaker {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "session_id", nullable = false)
-  private Session session;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "user_id", nullable = false)
-  private User user;
-
-  @Column
-  private String role;
-
-  public static SessionSpeaker of(Session session, User user) {
-    SessionSpeaker speaker = new SessionSpeaker();
-    speaker.session = session;
-    speaker.user = user;
-    return speaker;
-  }
-}
