+++ common/src/main/java/com/study/common/entity/sessionboard/Like.java
@@ -1,5 +1,6 @@
-package com.study.common.entity;
+package com.study.common.entity.sessionboard;
 
+import com.study.common.entity.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -10,30 +11,31 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.Table;
+import jakarta.persistence.UniqueConstraint;
 import java.time.OffsetDateTime;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "session_note")
+@Table(
+    name = "\"like\"",
+    uniqueConstraints = @UniqueConstraint(columnNames = {"user_id", "post_id"}))
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SessionNote {
+public class Like {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "session_id", nullable = false)
-  private Session session;
+  @JoinColumn(name = "user_id", nullable = false)
+  private User user;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "author_id", nullable = false)
-  private User author;
-
-  private String body;
+  @JoinColumn(name = "post_id", nullable = false)
+  private EventPost post;
 
   @Column(name = "created_at", nullable = false, updatable = false)
   private OffsetDateTime createdAt;
@@ -43,10 +45,10 @@ public class SessionNote {
     createdAt = OffsetDateTime.now();
   }
 
-  public static SessionNote of(Session session, User author) {
-    SessionNote note = new SessionNote();
-    note.session = session;
-    note.author = author;
-    return note;
+  public static Like of(User user, EventPost post) {
+    Like like = new Like();
+    like.user = user;
+    like.post = post;
+    return like;
   }
 }
