+++ common/src/main/java/com/study/common/entity/sessionboard/PostImage.java
@@ -1,0 +1,52 @@
+package com.study.common.entity.sessionboard;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.PrePersist;
+import jakarta.persistence.Table;
+import java.time.OffsetDateTime;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "post_image")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class PostImage {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "post_id", nullable = false)
+  private EventPost post;
+
+  @Column(nullable = false)
+  private String url;
+
+  @Column(name = "\"order\"", nullable = false)
+  private int order = 0;
+
+  @Column(name = "created_at", nullable = false, updatable = false)
+  private OffsetDateTime createdAt;
+
+  @PrePersist
+  void prePersist() {
+    createdAt = OffsetDateTime.now();
+  }
+
+  public static PostImage of(EventPost post, String url) {
+    PostImage image = new PostImage();
+    image.post = post;
+    image.url = url;
+    return image;
+  }
+}
