+++ common/src/main/java/com/study/common/entity/sessionboard/SessionSpeaker.java
@@ -1,5 +1,6 @@
-package com.study.common.entity;
+package com.study.common.entity.sessionboard;
 
+import com.study.common.entity.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -8,46 +9,39 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.PrePersist;
 import jakarta.persistence.Table;
 import jakarta.persistence.UniqueConstraint;
-import java.time.OffsetDateTime;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @Table(
-    name = "\"like\"",
-    uniqueConstraints = @UniqueConstraint(columnNames = {"user_id", "post_id"}))
+    name = "session_speaker",
+    uniqueConstraints = @UniqueConstraint(columnNames = {"session_id", "user_id"}))
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Like {
+public class SessionSpeaker {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
   @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "session_id", nullable = false)
+  private Session session;
+
+  @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "user_id", nullable = false)
   private User user;
 
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "post_id", nullable = false)
-  private EventPost post;
+  @Column
+  private String role;
 
-  @Column(name = "created_at", nullable = false, updatable = false)
-  private OffsetDateTime createdAt;
-
-  @PrePersist
-  void prePersist() {
-    createdAt = OffsetDateTime.now();
-  }
-
-  public static Like of(User user, EventPost post) {
-    Like like = new Like();
-    like.user = user;
-    like.post = post;
-    return like;
+  public static SessionSpeaker of(Session session, User user) {
+    SessionSpeaker speaker = new SessionSpeaker();
+    speaker.session = session;
+    speaker.user = user;
+    return speaker;
   }
 }
